nlp-module/requirements.txt
# Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Audio Processing
openai-whisper==20231117
edge-tts>=6.1.9
pydub==0.25.1
numpy==1.26.3

# LLM
groq>=0.4.0

# NLP Léger
scikit-learn==1.4.0
sentence-transformers==2.3.1
nltk==3.8.1

# Base de données
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# Développement
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
redis==5.0.1
pytest-cov==4.1.0
pytest-mock==3.12.0
fakeredis==2.21.0